+++ src/pages/myProfile/MyProfile.tsx
@@ -1,7 +1,5 @@
 import "./MyProfile.css";
-import CardProfileTidus from "../../components/CardProfileTidus/CardProfileTidus";
-import CardProfileTidusBack from "../../components/CardProfileTidus/CardProfileTidusBack";
-import CardProfileTidusFront from "../../components/CardProfileTidus/CardProfileTidusFront";
+import CardTidus from "../../components/CardTidus/CardTidus";
 import { useScreen } from "../../contexts/ScreenContext";
 import tidusProfile from "../../data/tidus.json";
 
@@ -12,13 +10,11 @@ function MyProfile() {
 	return (
 		<section className="profileContainer">
 			{isMobile ? (
-				<div className="tidusProfile">
-					<CardProfileTidus tidus={tidus} />
-				</div>
+				<CardTidus tidus={tidus} context={isMobile} />
 			) : (
 				<div className="tidusProfile">
-					<CardProfileTidusFront tidus={tidus} />
-					<CardProfileTidusBack tidus={tidus} />
+					<CardTidus tidus={tidus} context="front" />
+					<CardTidus tidus={tidus} context="back" />
 				</div>
 			)}
 		</section>
